refactor: 게임 진행 흐름에 속하지 않는 함수를 유틸 코드로 옮김
